Pretty big refactor for level spawning - Made a new level module for the spawn_level function and its helpers - Added functions to address repeated code for IntGrid spawning, metadata for intgrids, and metadata for tile layers
