+++ src/utils.rs
@@ -7,9 +7,8 @@ use crate::{
 };
 
 use crate::ldtk::*;
-use bevy::{prelude::*, sprite};
+use bevy::prelude::*;
 use bevy_ecs_tilemap::prelude::*;
-use thiserror::Error;
 
 use std::{collections::HashMap, hash::Hash};
 
@@ -316,67 +315,6 @@ pub fn sprite_bundle_from_entity_info(tileset: Option<&Handle<Image>>) -> Sprite
     SpriteBundle {
         texture: tileset,
         ..Default::default()
-    }
-}
-
-#[derive(Error, Debug)]
-pub(crate) enum BackgroundImageError {
-    #[error("background image handle not loaded into the image assets store")]
-    ImageNotLoaded,
-}
-
-pub(crate) fn background_image_sprite_sheet_bundle(
-    images: &Assets<Image>,
-    texture_atlases: &mut Assets<TextureAtlas>,
-    background_image_handle: &Handle<Image>,
-    background_position: &LevelBackgroundPosition,
-    level_height: i32,
-    transform_z: f32,
-) -> Result<SpriteSheetBundle, BackgroundImageError> {
-    if let Some(background_image) = images.get(background_image_handle) {
-        // We need to use a texture atlas to apply the correct crop to the image
-        let tile_size = Vec2::new(
-            background_image.texture_descriptor.size.width as f32,
-            background_image.texture_descriptor.size.height as f32,
-        );
-        let mut texture_atlas = TextureAtlas::new_empty(background_image_handle.clone(), tile_size);
-
-        let min = Vec2::new(
-            background_position.crop_rect[0],
-            background_position.crop_rect[1],
-        );
-
-        let size = Vec2::new(
-            background_position.crop_rect[2],
-            background_position.crop_rect[3],
-        );
-
-        let max = min + size;
-
-        let crop_rect = sprite::Rect { min, max };
-
-        texture_atlas.textures.push(crop_rect);
-
-        let texture_atlas_handle = texture_atlases.add(texture_atlas);
-
-        let scale = background_position.scale;
-
-        let scaled_size = size * scale;
-
-        let top_left_translation =
-            ldtk_pixel_coords_to_translation(background_position.top_left_px, level_height);
-
-        let center_translation =
-            top_left_translation + (Vec2::new(scaled_size.x, -scaled_size.y) / 2.);
-
-        Ok(SpriteSheetBundle {
-            texture_atlas: texture_atlas_handle,
-            transform: Transform::from_translation(center_translation.extend(transform_z))
-                .with_scale(scale.extend(1.)),
-            ..Default::default()
-        })
-    } else {
-        Err(BackgroundImageError::ImageNotLoaded)
     }
 }
